Reducción de cavalcade - cavalcade.py con todo lo de cava - mpris.py módulo - services/mpris.py
